decoupled config into _11ty folder
