+++ _11ty/minifyhtml.js
@@ -1,0 +1,19 @@
+'use strict'
+
+const htmlmin = require('html-minifier')
+
+module.exports = function minifyhtml (config) {
+  // Add HTML minification transform
+  config.addTransform('htmlmin', function (content, outputPath) {
+    if (outputPath.endsWith('.html')) {
+      const minified = htmlmin.minify(content, {
+        useShortDoctype: true,
+        removeComments: true,
+        collapseWhitespace: true
+      })
+      return minified
+    }
+
+    return content
+  })
+}
