+++ _11ty/toc.js
@@ -1,0 +1,10 @@
+'use strict'
+
+const pluginTOC = require('eleventy-plugin-toc')
+
+module.exports = function toc (config) {
+  config.addPlugin(pluginTOC, {
+    tags: ['h2', 'h3'],
+    wrapper: 'div'
+  })
+}
