+++ package.json
@@ -53,6 +53,7 @@
     "markdown-it": "^11.0.1",
     "markdown-it-anchor": "^6.0.0",
     "mini-css-extract-plugin": "^0.11.0",
+    "mkdirp": "^1.0.4",
     "npm-run-all": "^4.1.5",
     "purgecss": "^2.3.0",
     "sass": "^1.26.10",
